requirements.txt
flask
flask_session
flask_sqlalchemy
numpy
pandas
scikit-learn
requests
google.generativeai